move magic numbers into consts
